Move file cache into a separate selector
