+++ src/redux/csState.test.ts
@@ -6,22 +6,15 @@ import csReducer, {
   valueChanged,
   valuesChanged,
   unsubscribe,
-  fileChanged,
-  refreshFile,
   selectPvStates,
   pvStateComparator,
   selectDevice,
-  fileComparator,
-  selectFile,
   FullPvState,
   PvState,
   deviceComparator,
-  PvArrayResults,
-  findWidgetById,
-  fileDisplaySetGridLayout,
-  fileDisplaySetResponsiveLayout,
-  makeSelectWidgetPosition
+  PvArrayResults
 } from "./csState";
+import { findWidgetById } from "./slices/storeUtils";
 import {
   DType,
   dTypeGetAlarm,
@@ -41,7 +34,6 @@ import {
 } from "../testResources";
 import { WidgetDescription } from "../ui/widgets/createComponent";
 import { newAbsolutePosition } from "../types/position";
-import { ColorUtils } from "../types";
 
 const initialState: CsState = {
   valueCache: {
@@ -56,14 +48,6 @@ const initialState: CsState = {
   subscriptions: {},
   effectivePvNameMap: {},
   deviceCache: {},
-  fileCache: {
-    "mySecondFile.bob": {
-      fileId: "mySecondFile.bob",
-      id: "123",
-      type: "ellipse",
-      position: newAbsolutePosition("0", "0", "0", "0")
-    }
-  },
   pvwsSettings: {}
 };
 
@@ -242,38 +226,6 @@ test("handles initializers", (): void => {
   expect(state5.effectivePvNameMap["PV(1)"]).toEqual(undefined);
 });
 
-describe("FILE_CHANGED", (): void => {
-  test("csReducer adds file to fileCache", (): void => {
-    const contents: WidgetDescription = {
-      id: "123",
-      fileId: "AShapeFile",
-      type: "shape",
-      position: newAbsolutePosition("0", "0", "0", "0")
-    };
-    const fileName = "myfile.bob";
-    const action: ReturnType<typeof fileChanged> = {
-      type: "cs/fileChanged",
-      payload: { file: fileName, contents: contents }
-    };
-
-    const newState = csReducer(initialState, action);
-    expect(newState.fileCache[fileName]).toEqual(contents);
-  });
-});
-
-describe("REFRESH_FILE", (): void => {
-  test("csReducer deletes the file entry from fileCache", (): void => {
-    const fileName = "mySecondFile.bob";
-    const action: ReturnType<typeof refreshFile> = {
-      type: "cs/refreshFile",
-      payload: { file: fileName }
-    };
-
-    const newState = csReducer(initialState, action);
-    expect(newState.fileCache[fileName]).toBeUndefined();
-  });
-});
-
 describe("Selectors", () => {
   const pv1 = "pv1";
   const pv2 = "pv2";
@@ -291,7 +243,6 @@ describe("Selectors", () => {
     effectivePvNameMap: { pv1: "pv1", pv2: "pv3" },
     subscriptions: {},
     deviceCache: {},
-    fileCache: {},
     pvwsSettings: {}
   };
 
@@ -383,323 +334,6 @@ describe("Selectors", () => {
       const localState = { ...state, cs: { ...state.cs, deviceCache: {} } };
       expect(selectDevice(localState, "testDevice")).toBeUndefined();
     });
-  });
-
-  describe("fileComparator", (): void => {
-    it("returns false if string contents don't match", (): void => {
-      const contents1: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("0", "0", "0", "0")
-      };
-      const contents2: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("1", "0", "0", "0")
-      };
-      expect(fileComparator(contents1, contents2)).toBe(false);
-    });
-
-    it("returns false if number of keys changed", (): void => {
-      const contents1: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("0", "0", "0", "0"),
-        backgroundColor: ColorUtils.TRANSPARENT
-      };
-      const contents2: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("0", "0", "0", "0")
-      };
-
-      expect(fileComparator(contents1, contents2)).toBe(false);
-    });
-
-    it("returns true if matches", (): void => {
-      const contents: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("0", "0", "0", "0")
-      };
-      expect(fileComparator(contents, { ...contents })).toBe(true);
-    });
-  });
-
-  describe("selectFile", (): void => {
-    it("finds file in fileCache", (): void => {
-      const contents: WidgetDescription = {
-        id: "123",
-        fileId: "AShapeFile",
-        type: "shape",
-        position: newAbsolutePosition("0", "0", "0", "0")
-      };
-
-      const localState = {
-        ...state,
-        cs: {
-          ...state.cs,
-          fileCache: {
-            ...state.cs.fileCache,
-            "test.bob": contents
-          }
-        }
-      };
-
-      expect(selectFile(localState, "test.bob")).toEqual(contents);
-    });
-
-    it("returns undefined if device not in cache", (): void => {
-      const localState = { ...state, cs: { ...state.cs, fileCache: {} } };
-      expect(selectFile(localState, "test2.bob")).toBeUndefined();
-    });
-  });
-});
-
-describe("fileDisplaySetGridLayout", () => {
-  const baseDisplay = {
-    id: "display1",
-    type: "displayGridLayout",
-    fileId: "file",
-    children: [
-      {
-        id: "child1",
-        type: "shape",
-        position: newAbsolutePosition("10", "10", "20", "20")
-      }
-    ],
-    position: newAbsolutePosition("0", "0", "100", "100")
-  };
-
-  const initialState: CsState = {
-    valueCache: {},
-    globalMacros: {},
-    subscriptions: {},
-    effectivePvNameMap: {},
-    deviceCache: {},
-    fileCache: {
-      "file.bob": baseDisplay as any
-    },
-    pvwsSettings: {}
-  };
-
-  test("applies grid layout properties and normalises child positions", () => {
-    const action = fileDisplaySetGridLayout({
-      file: "file.bob",
-      displayId: "display1",
-      gridLayout: [{ i: "child1", x: 0, y: 0, w: 2, h: 2 }],
-      gridLayoutColumns: 12,
-      gridCellMargins: [5, 5],
-      gridCellHeight: 30,
-      gridCellDragEnabled: true,
-      gridCellResizeEnabled: false
-    });
-
-    const state = csReducer(initialState, action);
-    const display = state.fileCache["file.bob"];
-
-    expect(display.gridLayoutColumns).toBe(12);
-    expect(display.gridCellHeight).toBe(30);
-
-    const child = display?.children?.[0];
-    expect(child?.position).toMatchObject({
-      x: "0",
-      y: "0",
-      width: "100%",
-      height: "100%"
-    });
-  });
-
-  test("does nothing if display not found", () => {
-    const action = fileDisplaySetGridLayout({
-      file: "file.bob",
-      displayId: "missing",
-      gridLayout: [],
-      gridLayoutColumns: 12,
-      gridCellMargins: [0, 0],
-      gridCellHeight: 10,
-      gridCellDragEnabled: true,
-      gridCellResizeEnabled: true
-    });
-
-    const state = csReducer(initialState, action);
-    expect(state).toEqual(initialState);
-  });
-
-  test("does nothing if wrong display type", () => {
-    const badState: CsState = {
-      ...initialState,
-      fileCache: {
-        "file.bob": {
-          ...baseDisplay,
-          type: "shape" // wrong type
-        } as any
-      }
-    };
-
-    const action = fileDisplaySetGridLayout({
-      file: "file.bob",
-      displayId: "display1",
-      gridLayout: [],
-      gridLayoutColumns: 12,
-      gridCellMargins: [0, 0],
-      gridCellHeight: 10,
-      gridCellDragEnabled: true,
-      gridCellResizeEnabled: true
-    });
-
-    const state = csReducer(badState, action);
-    expect(state).toEqual(badState);
-  });
-});
-
-describe("fileDisplaySetResponsiveLayout", () => {
-  const baseDisplay = {
-    id: "display1",
-    type: "displayResponsive",
-    fileId: "file",
-    children: [
-      {
-        id: "child1",
-        type: "shape",
-        position: newAbsolutePosition("10", "10", "20", "20")
-      }
-    ],
-    position: newAbsolutePosition("0", "0", "100", "100")
-  };
-
-  const initialState: CsState = {
-    valueCache: {},
-    globalMacros: {},
-    subscriptions: {},
-    effectivePvNameMap: {},
-    deviceCache: {},
-    fileCache: {
-      "file.bob": baseDisplay as any
-    },
-    pvwsSettings: {}
-  };
-
-  test("applies responsive layout and updates child positions", () => {
-    const action = fileDisplaySetResponsiveLayout({
-      file: "file.bob",
-      displayId: "display1",
-      responsiveLayouts: { lg: [] },
-      responsiveColumns: { lg: 12 },
-      responsiveBreakpoints: { lg: 1200 },
-      gridCellMargins: [10, 10],
-      gridCellHeight: 50,
-      gridCellDragEnabled: true,
-      gridCellResizeEnabled: false
-    });
-
-    const state = csReducer(initialState, action);
-    const display = state.fileCache["file.bob"];
-
-    expect(display.responsiveColumns.lg).toBe(12);
-    expect(display.gridCellHeight).toBe(50);
-
-    // parent width forced to 100%
-    expect(display.position.width).toBe("100%");
-
-    const child = display?.children?.[0];
-    expect(child?.position).toMatchObject({
-      x: "0",
-      y: "0",
-      width: "100%",
-      height: "100%"
-    });
-  });
-
-  test("does nothing if display type is wrong", () => {
-    const badState = {
-      ...initialState,
-      fileCache: {
-        "file.bob": { ...baseDisplay, type: "shape" } as any
-      }
-    };
-
-    const action = fileDisplaySetResponsiveLayout({
-      file: "file.bob",
-      displayId: "display1",
-      responsiveLayouts: {},
-      responsiveColumns: {},
-      responsiveBreakpoints: {},
-      gridCellMargins: [0, 0],
-      gridCellHeight: 10,
-      gridCellDragEnabled: true,
-      gridCellResizeEnabled: true
-    });
-
-    const state = csReducer(badState, action);
-    expect(state).toEqual(badState);
-  });
-});
-
-describe("makeSelectWidgetPosition", () => {
-  const position = newAbsolutePosition("1", "2", "3", "4");
-
-  const file = {
-    id: "root",
-    type: "display",
-    fileId: "file",
-    children: [
-      {
-        id: "child1",
-        type: "shape",
-        position
-      }
-    ]
-  };
-
-  const state = createRootStoreState({
-    valueCache: {},
-    globalMacros: {},
-    subscriptions: {},
-    effectivePvNameMap: {},
-    deviceCache: {},
-    fileCache: {
-      "file.bob": file as any
-    },
-    pvwsSettings: {}
-  });
-
-  test("returns widget position when found", () => {
-    const selector = makeSelectWidgetPosition();
-
-    const result = selector(state, "file.bob", "child1");
-
-    expect(result).toEqual(position);
-  });
-
-  test("returns undefined if widget not found", () => {
-    const selector = makeSelectWidgetPosition();
-
-    const result = selector(state, "file.bob", "missing");
-
-    expect(result).toBeUndefined();
-  });
-
-  test("returns undefined if file not found", () => {
-    const selector = makeSelectWidgetPosition();
-
-    const result = selector(state, "missing.bob", "child1");
-
-    expect(result).toBeUndefined();
-  });
-
-  test("memoizes results (same inputs)", () => {
-    const selector = makeSelectWidgetPosition();
-
-    const result1 = selector(state, "file.bob", "child1");
-    const result2 = selector(state, "file.bob", "child1");
-
-    expect(result1).toBe(result2);
   });
 });
