+++ src/redux/csState.ts
@@ -1,11 +1,8 @@
 import log from "loglevel";
 import { MacroMap } from "../types/macros";
 import { mergeDType, DType } from "../types/dtypes";
-import { WidgetDescription } from "../ui/widgets/createComponent";
 import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { SubscriptionType } from "../connection";
-import { Breakpoints, Layout, ResponsiveLayouts } from "react-grid-layout";
-import { Position } from "../types/position";
 
 export const initialCsState: CsState = {
   valueCache: {},
@@ -13,7 +10,6 @@ export const initialCsState: CsState = {
   effectivePvNameMap: {},
   subscriptions: {},
   deviceCache: {},
-  fileCache: {},
   pvwsSettings: {}
 };
 
@@ -47,10 +43,6 @@ export interface DeviceCache {
   [deviceName: string]: DType;
 }
 
-export interface FileCache {
-  [fileName: string]: WidgetDescription;
-}
-
 export interface PvwsSettings {
   pvwsHost?: string;
 }
@@ -66,7 +58,6 @@ export interface CsState {
   globalMacros: MacroMap;
   subscriptions: Subscriptions;
   deviceCache: DeviceCache;
-  fileCache: FileCache;
   pvwsSettings: PvwsSettings;
 }
 
@@ -205,130 +196,10 @@ const csSlice = createSlice({
 
     queryDevice(_state, _action: PayloadAction<{ device: string }>) {
       // intentionally empty — handled by listener middleware
-    },
-
-    fileChanged(
-      state,
-      action: PayloadAction<{ file: string; contents: WidgetDescription }>
-    ) {
-      log.debug(action);
-      const { file, contents } = action.payload;
-      state.fileCache[file] = contents;
-    },
-
-    fileDisplaySetGridLayout(
-      state,
-      action: PayloadAction<{
-        file: string;
-        displayId: string;
-        gridLayout: Layout;
-        gridLayoutColumns: number;
-        gridCellMargins: [number, number];
-        gridCellHeight: number;
-        gridCellDragEnabled: boolean;
-        gridCellResizeEnabled: boolean;
-      }>
-    ) {
-      log.debug(action);
-      const {
-        file,
-        displayId,
-        gridLayout,
-        gridLayoutColumns,
-        gridCellMargins,
-        gridCellHeight,
-        gridCellDragEnabled,
-        gridCellResizeEnabled
-      } = action.payload;
-      const fileDescription = state.fileCache[file];
-      const display = findWidgetById([fileDescription], displayId);
-      if (!display || display.type !== "displayGridLayout") {
-        return;
-      }
-
-      // set grid layout props
-      display["gridLayout"] = gridLayout;
-      display["gridCellHeight"] = gridCellHeight;
-      display["gridCellMargins"] = gridCellMargins;
-      display["gridLayoutColumns"] = gridLayoutColumns;
-      display["gridCellDragEnabled"] = gridCellDragEnabled;
-      display["gridCellResizeEnabled"] = gridCellResizeEnabled;
-
-      // Set child positions, as position is now defined in the display layout
-      display?.children?.forEach(c => {
-        if (c.position) {
-          c.position["x"] = "0";
-          c.position["y"] = "0";
-          c.position["width"] = "100%";
-          c.position["height"] = "100%";
-        }
-      });
-    },
-
-    fileDisplaySetResponsiveLayout(
-      state,
-      action: PayloadAction<{
-        file: string;
-        displayId: string;
-        responsiveLayouts: ResponsiveLayouts<string>;
-        responsiveColumns: Breakpoints<string>;
-        responsiveBreakpoints: Breakpoints<string>;
-        gridCellMargins: [number, number];
-        gridCellHeight: number;
-        gridCellDragEnabled: boolean;
-        gridCellResizeEnabled: boolean;
-      }>
-    ) {
-      log.debug(action);
-      const {
-        file,
-        displayId,
-        responsiveLayouts,
-        responsiveColumns,
-        responsiveBreakpoints,
-        gridCellMargins,
-        gridCellHeight,
-        gridCellDragEnabled,
-        gridCellResizeEnabled
-      } = action.payload;
-      const fileDescription = state.fileCache[file];
-      const display = findWidgetById([fileDescription], displayId);
-      if (!display || display.type !== "displayResponsive") {
-        return;
-      }
-
-      // set responsive layout props
-      display["responsiveLayouts"] = responsiveLayouts;
-      display["responsiveColumns"] = responsiveColumns;
-      display["responsiveBreakpoints"] = responsiveBreakpoints;
-      display["gridCellHeight"] = gridCellHeight;
-      display["gridCellMargins"] = gridCellMargins;
-      display["gridCellDragEnabled"] = gridCellDragEnabled;
-      display["gridCellResizeEnabled"] = gridCellResizeEnabled;
-      if (display.position) {
-        display.position["width"] = "100%";
-      }
-
-      // Set child positions, as position is now defined in the display layout
-      display?.children?.forEach(c => {
-        if (c.position) {
-          c.position["x"] = "0";
-          c.position["y"] = "0";
-          c.position["width"] = "100%";
-          c.position["height"] = "100%";
-        }
-      });
-    },
-
-    refreshFile(state, action: PayloadAction<{ file: string }>) {
-      log.debug(action);
-      const { file } = action.payload;
-      delete state.fileCache[file];
     }
   },
   selectors: {
     selectDeviceCache: state => state.deviceCache,
-    selectFileCache: state => state.fileCache,
     selectEffectivePvNameMap: state => state.effectivePvNameMap,
     selectValueCache: state => state.valueCache,
     selectGlobalMacros: state => state.globalMacros,
@@ -346,16 +217,11 @@ export const {
   unsubscribe,
   writePv,
   deviceQueried,
-  queryDevice,
-  fileChanged,
-  fileDisplaySetGridLayout,
-  fileDisplaySetResponsiveLayout,
-  refreshFile
+  queryDevice
 } = csSlice.actions;
 export default csSlice.reducer;
 export const {
   selectDeviceCache,
-  selectFileCache,
   selectEffectivePvNameMap,
   selectValueCache,
   selectGlobalMacros,
@@ -371,28 +237,6 @@ export const selectDevice = createSelector(
   [selectDeviceCache, (_state, deviceId: string) => deviceId],
   (deviceCache, deviceId) => deviceCache[deviceId]
 );
-
-export const selectFile = createSelector(
-  [selectFileCache, (_state, fileId: string) => fileId],
-  (fileCache, fileId) => fileCache[fileId]
-);
-
-/**
- * This selector factory provides an alternative means for a widget to get its
- * position props. Generally the <Widget> wrapper should manage position and the
- * useContainerWidth hook should be used to get a measured width and height.
- * But in some rare situations these props need be be known before first render
- * and are not expected to change, for example through a react grid layout resize.
- * @returns A new selector that gets the widget position properties
- */
-export const makeSelectWidgetPosition = () =>
-  createSelector(
-    [selectFile, (_state: any, _fileId: string, widgetId: string) => widgetId],
-    (file, widgetId) =>
-      file
-        ? (findWidgetById([file], widgetId)?.position as Position | undefined)
-        : undefined
-  );
 
 export const selectPvStates = createSelector(
   [
@@ -411,27 +255,6 @@ export const selectPvStates = createSelector(
     return results;
   }
 );
-
-export const fileComparator = (
-  before: WidgetDescription,
-  after: WidgetDescription
-): boolean => {
-  if (!before || !after) {
-    return false;
-  }
-  if (Object.keys(before).length !== Object.keys(after).length) {
-    return false;
-  }
-  if (before.children?.length !== after.children?.length) {
-    return false;
-  }
-  // Can't compare objects directly because they are in different memory locations
-  // But we can compare strings
-  if (JSON.stringify(before) !== JSON.stringify(after)) {
-    return false;
-  }
-  return true;
-};
 
 /* Used for preventing re-rendering if the results are equivalent.
    Note that if the state for a particular PV hasn't changed, we will
@@ -474,28 +297,4 @@ export const deviceComparator = (before: DType, after: DType): boolean => {
     return false;
   }
   return true;
-};
-
-/**
- * Find an item in a tree by it's id.
- * @param tree The widget hierarchy
- * @param id The id of the widget to find.
- * @returns The matching widget description or undefined
- */
-export const findWidgetById = (
-  tree: WidgetDescription[] | undefined,
-  id: string
-): WidgetDescription | undefined => {
-  if (!Array.isArray(tree)) return undefined;
-
-  for (const node of tree) {
-    if (!node || typeof node !== "object") continue;
-
-    if (node.id === id) return node;
-
-    const found = findWidgetById(node.children, id);
-    if (found) return found;
-  }
-
-  return undefined;
 };
