+++ src/ui/hooks/useConnection.test.tsx
@@ -39,7 +39,6 @@ function getConnectionState(pvName: string, value: DType): CsState {
     globalMacros: {},
     effectivePvNameMap: {},
     deviceCache: {},
-    fileCache: {},
     pvwsSettings: {}
   };
 }
