+++ src/ui/hooks/useRules.test.tsx
@@ -38,7 +38,6 @@ function getCsState(value: DType): CsState {
     globalMacros: {},
     effectivePvNameMap: {},
     deviceCache: {},
-    fileCache: {},
     pvwsSettings: {}
   };
 }
